+++ examples/command_line/swift_c_module/BUILD
@@ -1,7 +1,7 @@
 load("@build_bazel_rules_swift//swift:swift.bzl", "swift_c_module")
 
 cc_library(
-    name = "lib_impl",
+    name = "c_lib",
     srcs = [
         "lib.c",
     ],
@@ -10,18 +10,12 @@ cc_library(
     ],
 )
 
-alias(
-    name = "lib2",
-    actual = "lib_impl",
-    visibility = ["//visibility:public"],
-)
-
 swift_c_module(
     name = "swift_c_module",
-    module_map = "lib2.modulemap",
+    module_map = "c_lib.modulemap",
     module_name = "SwiftCModule",
     visibility = ["//visibility:public"],
     deps = [
-        ":lib2",
+        ":c_lib",
     ],
 )
