+++ examples/command_line/swift_c_module/c_lib.c
@@ -1,4 +1,4 @@
-#include "lib.h"
+#include "c_lib.h"
 
 int answer() {
     return 42;
