+++ examples/command_line/swift_c_module/c_lib.h
