+++ test/fixtures/command_line/bwb.xcodeproj/xcshareddata/xcschemes/lib_impl.xcscheme
@@ -5,6 +5,24 @@
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
+      <PreActions>
+         <ExecutionAction
+            ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
+            <ActionContent
+               title = "Set Bazel Build Output Groups"
+               scriptText = "mkdir -p &quot;${BAZEL_BUILD_OUTPUT_GROUPS_FILE%/*}&quot;&#10;echo &quot;b $BAZEL_TARGET_ID&quot; &gt; &quot;$BAZEL_BUILD_OUTPUT_GROUPS_FILE&quot;&#10;">
+               <EnvironmentBuildable>
+                  <BuildableReference
+                     BuildableIdentifier = "primary"
+                     BlueprintIdentifier = "BC691472A2BC47F8E1D5D637"
+                     BuildableName = "lib_impl"
+                     BlueprintName = "lib_impl"
+                     ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
+                  </BuildableReference>
+               </EnvironmentBuildable>
+            </ActionContent>
+         </ExecutionAction>
+      </PreActions>
       <BuildActionEntries>
          <BuildActionEntry
             buildForTesting = "YES"
@@ -14,10 +32,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2C1A4EFEDCA8D354A077D8EE"
-               BuildableName = "//examples/command_line/lib:lib_impl"
-               BlueprintName = "//examples/command_line/lib:lib_impl"
-               ReferencedContainer = "container:test/fixtures/command_line/bwx.xcodeproj">
+               BlueprintIdentifier = "BC691472A2BC47F8E1D5D637"
+               BuildableName = "lib_impl"
+               BlueprintName = "lib_impl"
+               ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
