+++ test/fixtures/command_line/bwb_spec.json
@@ -27,7 +27,7 @@
             "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/lib/lib_impl.swift.modulemap",
             "t": "g"
         },
-        "examples/command_line/swift_c_module/lib2.modulemap",
+        "examples/command_line/swift_c_module/c_lib.modulemap",
         {
             "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/external/examples_command_line_external/Library.swift.modulemap",
             "t": "g"
@@ -139,7 +139,7 @@
                         "t": "g"
                     },
                     {
-                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/libc_lib.a",
                         "t": "g"
                     },
                     {
@@ -219,7 +219,7 @@
                     "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/lib/lib_impl.swift.modulemap",
                     "t": "g"
                 },
-                "examples/command_line/swift_c_module/lib2.modulemap",
+                "examples/command_line/swift_c_module/c_lib.modulemap",
                 {
                     "_": "examples_command_line_external/ExternalFramework.framework/Modules/module.modulemap",
                     "t": "e"
@@ -426,7 +426,7 @@
             },
             "configuration": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
             "dependencies": [
-                "//examples/command_line/swift_c_module:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
+                "//examples/command_line/swift_c_module:c_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
                 "//examples/command_line/lib:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
                 "//examples/command_line/lib:private_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630"
             ],
@@ -445,7 +445,7 @@
                     "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/lib/lib_impl.swift.modulemap",
                     "t": "g"
                 },
-                "examples/command_line/swift_c_module/lib2.modulemap",
+                "examples/command_line/swift_c_module/c_lib.modulemap",
                 {
                     "_": "examples_command_line_external/ExternalFramework.framework/Modules/module.modulemap",
                     "t": "e"
@@ -610,7 +610,7 @@
             "test_host": null,
             "watch_application": null
         },
-        "//examples/command_line/swift_c_module:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
+        "//examples/command_line/swift_c_module:c_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630",
         {
             "build_settings": {
                 "CLANG_CXX_LANGUAGE_STANDARD": "c++0x",
@@ -663,10 +663,10 @@
                 ]
             },
             "is_swift": false,
-            "label": "//examples/command_line/swift_c_module:lib_impl",
+            "label": "//examples/command_line/swift_c_module:c_lib",
             "linker_inputs": {},
             "modulemaps": [],
-            "name": "lib_impl",
+            "name": "c_lib",
             "outputs": {},
             "package_bin_dir": "bazel-out/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module",
             "platform": {
@@ -676,9 +676,9 @@
                 "os": "macos"
             },
             "product": {
-                "name": "lib_impl",
+                "name": "c_lib",
                 "path": {
-                    "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                    "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/libc_lib.a",
                     "t": "g"
                 },
                 "type": "com.apple.product-type.library.static"
@@ -765,7 +765,7 @@
                         "t": "g"
                     },
                     {
-                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-0139d977e630/bin/examples/command_line/swift_c_module/libc_lib.a",
                         "t": "g"
                     },
                     {
