+++ test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/targets/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/examples/command_line/tool/tool.LinkFileList
@@ -1,5 +1,5 @@
 test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/gen_dir/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/liblib_swift.a
-test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/gen_dir/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/liblib_impl.a
+test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/gen_dir/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/libc_lib.a
 test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/gen_dir/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/libprivate_lib.a
 test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/gen_dir/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/liblib_impl.a
 test/fixtures/command_line/bwx.xcodeproj/rules_xcodeproj/links/external/examples_command_line_external/ImportableLibrary/libImportableLibrary.a
