+++ test/fixtures/command_line/bwx.xcodeproj/xcshareddata/xcschemes/c_lib.xcscheme
@@ -5,24 +5,6 @@
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
-      <PreActions>
-         <ExecutionAction
-            ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
-            <ActionContent
-               title = "Set Bazel Build Output Groups"
-               scriptText = "mkdir -p &quot;${BAZEL_BUILD_OUTPUT_GROUPS_FILE%/*}&quot;&#10;echo &quot;b $BAZEL_TARGET_ID&quot; &gt; &quot;$BAZEL_BUILD_OUTPUT_GROUPS_FILE&quot;&#10;">
-               <EnvironmentBuildable>
-                  <BuildableReference
-                     BuildableIdentifier = "primary"
-                     BlueprintIdentifier = "E6CDBA8D28089F204E6BBEC9"
-                     BuildableName = "//examples/command_line/swift_c_module:lib_impl"
-                     BlueprintName = "//examples/command_line/swift_c_module:lib_impl"
-                     ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
-                  </BuildableReference>
-               </EnvironmentBuildable>
-            </ActionContent>
-         </ExecutionAction>
-      </PreActions>
       <BuildActionEntries>
          <BuildActionEntry
             buildForTesting = "YES"
@@ -32,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "E6CDBA8D28089F204E6BBEC9"
-               BuildableName = "//examples/command_line/swift_c_module:lib_impl"
-               BlueprintName = "//examples/command_line/swift_c_module:lib_impl"
-               ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
+               BlueprintIdentifier = "F38BE7F734335EF255F553F2"
+               BuildableName = "c_lib"
+               BlueprintName = "c_lib"
+               ReferencedContainer = "container:test/fixtures/command_line/bwx.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
