+++ test/fixtures/command_line/bwx.xcodeproj/xcshareddata/xcschemes/lib_impl.xcscheme
@@ -5,24 +5,6 @@
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
-      <PreActions>
-         <ExecutionAction
-            ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
-            <ActionContent
-               title = "Set Bazel Build Output Groups"
-               scriptText = "mkdir -p &quot;${BAZEL_BUILD_OUTPUT_GROUPS_FILE%/*}&quot;&#10;echo &quot;b $BAZEL_TARGET_ID&quot; &gt; &quot;$BAZEL_BUILD_OUTPUT_GROUPS_FILE&quot;&#10;">
-               <EnvironmentBuildable>
-                  <BuildableReference
-                     BuildableIdentifier = "primary"
-                     BlueprintIdentifier = "095FE5B4B5B509FE4AB36B0F"
-                     BuildableName = "//examples/command_line/lib:lib_impl"
-                     BlueprintName = "//examples/command_line/lib:lib_impl"
-                     ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
-                  </BuildableReference>
-               </EnvironmentBuildable>
-            </ActionContent>
-         </ExecutionAction>
-      </PreActions>
       <BuildActionEntries>
          <BuildActionEntry
             buildForTesting = "YES"
@@ -32,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "095FE5B4B5B509FE4AB36B0F"
-               BuildableName = "//examples/command_line/lib:lib_impl"
-               BlueprintName = "//examples/command_line/lib:lib_impl"
-               ReferencedContainer = "container:test/fixtures/command_line/bwb.xcodeproj">
+               BlueprintIdentifier = "8329F08BE01D4CC32B819CE9"
+               BuildableName = "lib_impl"
+               BlueprintName = "lib_impl"
+               ReferencedContainer = "container:test/fixtures/command_line/bwx.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
