+++ test/fixtures/command_line/bwx_spec.json
@@ -27,7 +27,7 @@
             "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/lib_impl.swift.modulemap",
             "t": "g"
         },
-        "examples/command_line/swift_c_module/lib2.modulemap",
+        "examples/command_line/swift_c_module/c_lib.modulemap",
         {
             "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/external/examples_command_line_external/Library.swift.modulemap",
             "t": "g"
@@ -139,7 +139,7 @@
                         "t": "g"
                     },
                     {
-                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/libc_lib.a",
                         "t": "g"
                     },
                     {
@@ -214,7 +214,7 @@
                     "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/lib_impl.swift.modulemap",
                     "t": "g"
                 },
-                "examples/command_line/swift_c_module/lib2.modulemap",
+                "examples/command_line/swift_c_module/c_lib.modulemap",
                 {
                     "_": "examples_command_line_external/ExternalFramework.framework/Modules/module.modulemap",
                     "t": "e"
@@ -406,7 +406,7 @@
             },
             "configuration": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
             "dependencies": [
-                "//examples/command_line/swift_c_module:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
+                "//examples/command_line/swift_c_module:c_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
                 "//examples/command_line/lib:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
                 "//examples/command_line/lib:private_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60"
             ],
@@ -425,7 +425,7 @@
                     "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/lib/lib_impl.swift.modulemap",
                     "t": "g"
                 },
-                "examples/command_line/swift_c_module/lib2.modulemap",
+                "examples/command_line/swift_c_module/c_lib.modulemap",
                 {
                     "_": "examples_command_line_external/ExternalFramework.framework/Modules/module.modulemap",
                     "t": "e"
@@ -571,7 +571,7 @@
             "test_host": null,
             "watch_application": null
         },
-        "//examples/command_line/swift_c_module:lib_impl macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
+        "//examples/command_line/swift_c_module:c_lib macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60",
         {
             "build_settings": {
                 "CLANG_CXX_LANGUAGE_STANDARD": "c++0x",
@@ -624,10 +624,10 @@
                 ]
             },
             "is_swift": false,
-            "label": "//examples/command_line/swift_c_module:lib_impl",
+            "label": "//examples/command_line/swift_c_module:c_lib",
             "linker_inputs": {},
             "modulemaps": [],
-            "name": "lib_impl",
+            "name": "c_lib",
             "outputs": {},
             "package_bin_dir": "bazel-out/macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module",
             "platform": {
@@ -637,9 +637,9 @@
                 "os": "macos"
             },
             "product": {
-                "name": "lib_impl",
+                "name": "c_lib",
                 "path": {
-                    "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                    "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/libc_lib.a",
                     "t": "g"
                 },
                 "type": "com.apple.product-type.library.static"
@@ -726,7 +726,7 @@
                         "t": "g"
                     },
                     {
-                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/liblib_impl.a",
+                        "_": "macos-x86_64-min11.0-applebin_macos-darwin_x86_64-dbg-ST-59a3e2f8ef60/bin/examples/command_line/swift_c_module/libc_lib.a",
                         "t": "g"
                     },
                     {
